+++ fuse/__init__.py
@@ -1,4 +1,7 @@
-__version__ = "1.1.0"
+__version__ = "1.2.2"
+
+from . import dtypes
+from .dtypes import *
 
 from . import plugins
 from .plugins import *