+++ tests/test_input.py
@@ -42,8 +42,10 @@ class TestInput(unittest.TestCase):
             }
         )
         g4_loaded = test_context.get_array(self.run_number, "geant4_interactions")
-        loaded_event_count = len(np.unique(g4_loaded["evtid"]))
-        assert loaded_event_count == 52, f"Expecting 52 events, but got {loaded_event_count} events"
+        loaded_event_count = len(np.unique(g4_loaded["eventid"]))
+        self.assertTrue(
+            loaded_event_count == 52, f"Expecting 52 events, but got {loaded_event_count} events"
+        )
 
     @timeout_decorator.timeout(TIMEOUT, exception_message="LoadHalf timed out")
     def test_load_half(self):
@@ -57,9 +59,11 @@ class TestInput(unittest.TestCase):
             }
         )
         g4_loaded = test_context.get_array(self.run_number, "geant4_interactions")
-        loaded_event_count = len(np.unique(g4_loaded["evtid"]))
+        loaded_event_count = len(np.unique(g4_loaded["eventid"]))
 
-        assert loaded_event_count == 26, f"Expecting 26 events, but got {loaded_event_count} events"
+        self.assertTrue(
+            loaded_event_count == 26, f"Expecting 26 events, but got {loaded_event_count} events"
+        )
 
     @timeout_decorator.timeout(TIMEOUT, exception_message="LoadEventIDAll timed out")
     def test_load_eventid_all(self):
@@ -68,8 +72,10 @@ class TestInput(unittest.TestCase):
             {"path": self.temp_dir.name, "file_name": test_root_file_name, "cut_by_eventid": True}
         )
         g4_loaded = test_context.get_array(self.run_number, "geant4_interactions")
-        loaded_event_count = len(np.unique(g4_loaded["evtid"]))
-        assert loaded_event_count == 52, f"Expecting 52 events, but got {loaded_event_count} events"
+        loaded_event_count = len(np.unique(g4_loaded["eventid"]))
+        self.assertTrue(
+            loaded_event_count == 52, f"Expecting 52 events, but got {loaded_event_count} events"
+        )
 
     @timeout_decorator.timeout(TIMEOUT, exception_message="LoadEventIDHalf timed out")
     def test_load_eventid_half(self):
@@ -84,8 +90,10 @@ class TestInput(unittest.TestCase):
             }
         )
         g4_loaded = test_context.get_array(self.run_number, "geant4_interactions")
-        loaded_event_count = len(np.unique(g4_loaded["evtid"]))
-        assert loaded_event_count == 23, f"Expecting 23 events, but got {loaded_event_count} events"
+        loaded_event_count = len(np.unique(g4_loaded["eventid"]))
+        self.assertTrue(
+            loaded_event_count == 23, f"Expecting 23 events, but got {loaded_event_count} events"
+        )
 
     @timeout_decorator.timeout(TIMEOUT, exception_message="InvalidArgs0 timed out")
     def test_invalid_args_0(self):
@@ -99,11 +107,8 @@ class TestInput(unittest.TestCase):
                 "entry_stop": 15,
             }
         )
-        try:
+        with self.assertRaises(ValueError):
             test_context.make(self.run_number, "geant4_interactions")
-        except ValueError:
-            return
-        raise RuntimeError("entry_start >= entry_stop does not raise an exception!")
 
     @timeout_decorator.timeout(TIMEOUT, exception_message="InvalidArgs1 timed out")
     def test_invalid_args_1(self):
@@ -117,13 +122,8 @@ class TestInput(unittest.TestCase):
                 "entry_stop": 91,
             }
         )
-        try:
+        with self.assertRaises(ValueError):
             test_context.make(self.run_number, "geant4_interactions")
-        except ValueError:
-            return
-        raise RuntimeError(
-            "An out-of-range entry_start without cut_by_eventid does not raise an exception!"
-        )
 
     @timeout_decorator.timeout(TIMEOUT, exception_message="InvalidArgs2 timed out")
     def test_invalid_args_2(self):
@@ -137,13 +137,8 @@ class TestInput(unittest.TestCase):
                 "entry_stop": -1,
             }
         )
-        try:
+        with self.assertRaises(ValueError):
             test_context.make(self.run_number, "geant4_interactions")
-        except ValueError:
-            return
-        raise RuntimeError(
-            "An out-of-range entry_stop without cut_by_eventid does not raise an exception!"
-        )
 
     @timeout_decorator.timeout(TIMEOUT, exception_message="InvalidArgs3 timed out")
     def test_invalid_args_3(self):
@@ -157,13 +152,8 @@ class TestInput(unittest.TestCase):
                 "entry_stop": -1,
             }
         )
-        try:
+        with self.assertRaises(ValueError):
             test_context.make(self.run_number, "geant4_interactions")
-        except ValueError:
-            return
-        raise RuntimeError(
-            "An out-of-range entry_stop with cut_by_eventid does not raise an exception!"
-        )
 
     @timeout_decorator.timeout(TIMEOUT, exception_message="InvalidArgs4 timed out")
     def test_invalid_args_4(self):
@@ -177,13 +167,8 @@ class TestInput(unittest.TestCase):
                 "entry_stop": 103,
             }
         )
-        try:
+        with self.assertRaises(ValueError):
             test_context.make(self.run_number, "geant4_interactions")
-        except ValueError:
-            return
-        raise RuntimeError(
-            "An out-of-range entry_start with cut_by_eventid does not raise an exception!"
-        )
 
     @timeout_decorator.timeout(TIMEOUT, exception_message="InvalidArgs5 timed out")
     def test_invalid_args_5(self):
@@ -197,13 +182,8 @@ class TestInput(unittest.TestCase):
                 "entry_stop": 11,
             }
         )
-        try:
+        with self.assertRaises(ValueError):
             test_context.make(self.run_number, "geant4_interactions")
-        except ValueError:
-            return
-        raise RuntimeError(
-            "Selecting an empty eventid range with cut_by_eventid does not raise an exception!"
-        )
 
 
 if __name__ == "__main__":