+++ HISTORY.md
@@ -1,3 +1,34 @@
+1.2.2 / 2024-05-03
+------------------
+* Add Zenodo batch (f08d6ce)
+* pre-commit autoupdate (#205)
+* No photons from electrons outside (#200)
+* Rename s1 photon hits output (#206)
+* Change g4_fields dtype for t (#209)
+* Add x_truth, y_truth, z_truth (#202)
+* Add tight coincidence and peak tagging in `ClusterTagging` (#210)
+
+1.2.1 / 2024-04-24
+------------------
+* pre-commit autoupdate (#191)
+* Fix the `data_kind` of `RecordsTruth` (#194)
+* Test full chain in multiple chunks and clean up input tests (#184)
+* Add a dtypes manager (#195)
+* Add pull_request_template.md (#197)
+* Scale S2 pattern map to correct S2 AFT (#196)
+
+1.2.0 / 2024-04-11
+------------------
+* Specify lxml version (#185)
+* Avoid duplicate code in downchunking (#180)
+* Exclude PMT AP photons from raw_area_truth calculation (#176)
+* Optional random seed as user input (#179)
+* pre-commit autoupdate (#186)
+* Automatic Plugin Documentation (#187)
+* Embellish some codes (#188)
+* Separate above- and below-gate drift of electrons (#178)
+* Add photoionization plugins (#52)
+
 1.1.0 / 2024-03-20
 ------------------
 * Cleanup and refactor configurations (#163)